requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.4.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
autopep8==2.0.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==3.3.1
blinker==1.4
blis==0.7.9
Brlapi==0.7.0
cachetools==4.2.2
catalogue==2.0.8
certifi==2022.9.24
cffi==1.14.6
chardet==3.0.4
clang==5.0
click==8.0.1
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
commonmark==0.9.1
confection==0.0.4
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.7
datasets==2.6.1
dbus-python==1.2.16
debugpy==1.3.0
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
dill==0.3.5.1
discord.py==1.7.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.4
dnspython==2.2.1
duplicity==0.8.12.0
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
fastapi==0.65.2
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
fsspec==2022.10.0
future==0.18.2
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.38.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
httplib2==0.14.0
httptools==0.2.0
huggingface-hub==0.10.1
humanfriendly==9.2
idna==2.8
imageio==2.9.0
img2pdf==0.4.1
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.18.0
jina==2.0.0rc11.dev39
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
librosa==0.8.1
llvmlite==0.37.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
lz4==3.1.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.5
MouseInfo==0.1.3
multidict==5.1.0
multiprocess==0.70.13
murmurhash==1.0.9
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
nltk==3.7
notebook==6.4.0
notify2==0.3
numba==0.54.0
numpy==1.22.4
oauthlib==3.1.0
ocrmypdf==12.3.0
opencv-python==4.5.3.56
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==20.9
pandas==1.5.1
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pathy==0.10.1
pdfminer.six==20201018
pexpect==4.6.0
pickleshare==0.7.5
pikepdf==2.15.0
Pillow==8.3.1
pluggy==0.13.1
pocketsphinx==0.1.15
pooch==1.5.1
preshed==3.0.8
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py==1.11.0
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.9.1
pycparser==2.20
pycups==1.9.73
pydantic==1.8.2
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.2.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==3.0.0b2
PyPDF2==1.26.0
pyperclip==1.8.2
PyQt5==5.14.1
PyRect==0.1.4
pyRFC3339==1.1
pyrsistent==0.18.0
PyScreeze==0.1.27
pysentimiento==0.5.2
PySocks==1.7.1
pytest==7.1.3
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.18.0
python-multipart==0.0.6
python3-xlib==0.15
PyTweening==1.0.3
pytz==2022.6
PyWavelets==1.1.1
pywhatkit==5.4
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
rdflib==6.0.1
regex==2022.9.13
reportlab==3.5.68
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resampy==0.2.2
responses==0.18.0
rich==12.6.0
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
SecretStorage==2.3.1
selenium==4.5.0
Send2Trash==1.7.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
simplejson==3.16.0
sip==4.19.21
six==1.15.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.8.1
sqlparse==0.4.1
srsly==2.4.6
starlette==0.14.2
systemd-python==234
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.10.1
tesserocr==2.5.2
testpath==0.5.0
textblob==0.17.1
thinc==8.1.9
threadpoolctl==2.2.0
tifffile==2021.7.2
tokenizers==0.13.1
tomli==2.0.1
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.64.1
traitlets==5.0.5
transformers==4.24.0
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
uvicorn==0.14.0
uvloop==0.15.2
wadllib==1.3.3
wasabi==1.1.1
watchgod==0.7
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.1
wikipedia==1.4.0
wrapt==1.12.1
wsproto==1.2.0
xkit==0.0.0
xxhash==3.1.0
yarl==1.6.3
youtube-dl==2021.6.6
